Support Sidekiq, Resque, or the basic Rails 4 queue
In preparation of Rails 4.0, support for Rails::Queueing is now here.
Additionally, a worker for Sidekiq has been included. This means that
three major queueing systems will now be supported: Resque, Sidekiq, and
Rails::Queueing. Rails::Queueing will be the default if no other queue
system is bundled. If Sidekiq or Resque are bundled, they will be
preferred. I am assuming that nobody will bundle both but, if for some
reason they do, Sidekiq is preferred.

Signed-off-by: David Celis <david@davidcelis.com>