Lazily cache the counts.
I think the whole count caching is premature optimization, but let's
roll with it for now. Main change here is to not do the cache when the
code is evaluated, but when the method is executed (lazily), because the
cache will run as soon as the code is required. This will fail if we're
for instance running `db:schema:load`, because the db doesn't exist yet.